+++ src/clients/catalog.ts
@@ -9,13 +9,13 @@ import { getAuthToken } from '../utils/authToken'
 import { createTracing } from '../utils/tracing'
 import { checkSellerInformation } from '../utils/seller'
 import { AuthMethod } from '../typings/tokens'
-import { OrderFormConfiguration } from '../typings/orderForm'
-import { Seller } from '../typings/catalog'
+import { GetSkuResponse, Seller } from '../typings/catalog'
 
 const baseURL = '/api/catalog'
 const baseURLLegacy = '/api/catalog_system'
 
 const routes = {
+  productsAndSkus: `${baseURLLegacy}/pvt/products/GetProductAndSkuIds`,
   getSkuById: (skuId: string) => `${baseURL}/pvt/stockkeepingunit/${skuId}`,
   changeNotification: (sellerId: string, skuId: string) =>
     `${baseURLLegacy}/pvt/skuseller/changenotification/${sellerId}/${skuId}`,
@@ -30,6 +30,24 @@ export class Catalog extends JanusClient {
     })
   }
 
+  public getProductsAndSkus(
+    authMethod: AuthMethod = 'AUTH_TOKEN',
+    tracingConfig?: RequestTracingConfig
+  ) {
+    const metric = 'catalog-getProductsAndSkus'
+    const token = getAuthToken(this.context, authMethod)
+
+    return this.http.get(routes.productsAndSkus, {
+      headers: token
+        ? {
+            VtexIdclientAutCookie: token,
+          }
+        : {},
+      metric,
+      tracing: createTracing(metric, tracingConfig),
+    })
+  }
+
   public getSkuById(
     skuId: string,
     authMethod: AuthMethod = 'AUTH_TOKEN',
@@ -38,7 +56,7 @@ export class Catalog extends JanusClient {
     const metric = 'catalog-getSkuMetric'
     const token = getAuthToken(this.context, authMethod)
 
-    return this.http.get<OrderFormConfiguration>(routes.getSkuById(skuId), {
+    return this.http.get<GetSkuResponse>(routes.getSkuById(skuId), {
       headers: token
         ? {
             VtexIdclientAutCookie: token,