+++ src/typings/affiliate.ts
@@ -1,6 +1,7 @@
 export interface AffiliateInput {
   name: string
   id: string
+  followUpEmail: string
   salesChannelId: string
   searchEndpoint: string
 }