+++ CHANGELOG.md
@@ -6,9 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.1.2-beta] - 2020-12-22
+## [2.1.0] - 2020-12-07
+### Fixed
+- Response typing of Catalog method
+### Added
+- New methods to Catalog, OMS and Logistics clients
 
-## [1.1.1-beta] - 2020-12-22
+## [2.0.0] - 2020-12-03
+### Fixed
+- Removes mocked field on a Affiliate client's method
 
 ## [1.1.0] - 2020-12-02